feat(log change) - tinggal finishing UI

+++ Ardita.Repositories/Classess/PageDetailRepository.cs
@@ -1,4 +1,5 @@
-﻿using Ardita.Models.DbModels;
+﻿using Ardita.Extensions;
+using Ardita.Models.DbModels;
 using Ardita.Repositories.Interfaces;
 using Microsoft.EntityFrameworkCore;
 
@@ -7,9 +8,11 @@ namespace Ardita.Repositories.Classess
     public class PageDetailRepository : IPageDetailRepository
     {
         private readonly BksArditaDevContext _context;
-        public PageDetailRepository(BksArditaDevContext context)
+        private readonly ILogChangesRepository _logChangesRepository;
+        public PageDetailRepository(BksArditaDevContext context, ILogChangesRepository logChangesRepository)
         {
             _context = context;
+            _logChangesRepository = logChangesRepository;
         }
         public async Task<int> Delete(MstPageDetail model)
         {
@@ -19,6 +22,16 @@ namespace Ardita.Repositories.Classess
             {
                 _context.MstPageDetails.Remove(model);
                 result = await _context.SaveChangesAsync();
+
+                //Log
+                if (result > 0)
+                {
+                    try
+                    {
+                        await _logChangesRepository.CreateLog<MstPageDetail>(GlobalConst.Delete, (Guid)model.CreatedBy!, new List<MstPageDetail> { model }, new List<MstPageDetail> {  });
+                    }
+                    catch (Exception ex) { }
+                }
             }
             return result;
         }
@@ -28,8 +41,22 @@ namespace Ardita.Repositories.Classess
             int result = 0;
             if (id != null)
             {
-                _context.Database.ExecuteSqlRaw($" delete from dbo.MST_PAGE_DETAIL where page_id='{id}'");
-                result = await _context.SaveChangesAsync();
+                var dataOld = await _context.MstPageDetails.Where(x => x.PageId == id).ToListAsync();
+                if(dataOld != null)
+                {
+                    _context.Database.ExecuteSqlRaw($" delete from dbo.MST_PAGE_DETAIL where page_id='{id}'");
+                    result = await _context.SaveChangesAsync();
+
+                    //Log
+                    if (result > 0)
+                    {
+                        try
+                        {
+                            await _logChangesRepository.CreateLog<MstPageDetail>(GlobalConst.Delete, (Guid)dataOld.FirstOrDefault()!.CreatedBy!, dataOld, new List<MstPageDetail> { });
+                        }
+                        catch (Exception ex) { }
+                    }
+                }
             }
             return result;
         }
@@ -68,6 +95,16 @@ namespace Ardita.Repositories.Classess
                     _context.MstPageDetails.Add(model);
                     result = await _context.SaveChangesAsync();
                 }
+
+                //Log
+                if (result > 0)
+                {
+                    try
+                    {
+                        await _logChangesRepository.CreateLog<MstPageDetail>(GlobalConst.New, (Guid)model.CreatedBy!, new List<MstPageDetail> { }, new List<MstPageDetail> { model });
+                    }
+                    catch (Exception ex) { }
+                }
             }
             return result;
         }
@@ -78,11 +115,21 @@ namespace Ardita.Repositories.Classess
 
             if (model != null && model.PageDetailId != Guid.Empty)
             {
-                var data = await _context.MstPageDetails.AsNoTracking().Where(x => x.PageDetailId == model.PageDetailId).ToListAsync();
+                var data = await _context.MstPageDetails.AsNoTracking().Where(x => x.PageDetailId == model.PageDetailId).FirstOrDefaultAsync();
                 if (data != null)
                 {
                     _context.Update(model);
                     result = await _context.SaveChangesAsync();
+
+                    //Log
+                    if (result > 0)
+                    {
+                        try
+                        {
+                            await _logChangesRepository.CreateLog<MstPageDetail>(GlobalConst.Update, (Guid)model.CreatedBy!, new List<MstPageDetail> { data }, new List<MstPageDetail> { model });
+                        }
+                        catch (Exception ex) { }
+                    }
                 }
             }
             return result;
